requirements.txt
flask==3.0.0
flask-cors==4.0.0
web3==6.11.3
psycopg2-binary==2.9.9
